Refactor to a function, Created instance structure
Moved the core of the instance gathering code to a function. This
will enable us to call it several times for different regions.

Created an instance structure for just the fields we want, populate
that structure as we step through each instance in a reservation.

Added some logic to determine single letter running state, and if
the instance is SPOT or not.

Thanks to @iamtommythorn for the ".clone().unwrap_or_default()"